Sort commits using isort
Also added lint test to ensure they stay sorted.
