+++ datasette/__init__.py
@@ -1,3 +1,4 @@
-from datasette.version import __version_info__, __version__  # noqa
+from datasette.version import __version__, __version_info__  # noqa
+
 from .hookspecs import hookimpl  # noqa
 from .hookspecs import hookspec  # noqa
