+++ datasette/plugins.py
@@ -1,6 +1,8 @@
 import importlib
-import pluggy
 import sys
+
+import pluggy
+
 from . import hookspecs
 
 DEFAULT_PLUGINS = (
