+++ datasette/publish/common.py
@@ -1,7 +1,9 @@
-from ..utils import StaticMount
-import click
 import shutil
 import sys
+
+import click
+
+from ..utils import StaticMount
 
 
 def add_common_publish_arguments_and_options(subcommand):
