+++ tests/test_publish_now.py
@@ -1,7 +1,9 @@
-from click.testing import CliRunner
-from datasette import cli
-from unittest import mock
 import subprocess
+from unittest import mock
+
+from click.testing import CliRunner
+
+from datasette import cli
 
 
 @mock.patch("shutil.which")
