+++ versioneer.py
@@ -277,16 +277,18 @@ https://creativecommons.org/publicdomain/zero/1.0/ .
 """
 
 from __future__ import print_function
-try:
-    import configparser
-except ImportError:
-    import ConfigParser as configparser
+
 import errno
 import json
 import os
 import re
 import subprocess
 import sys
+
+try:
+    import configparser
+except ImportError:
+    import ConfigParser as configparser
 
 
 class VersioneerConfig:
